Moved archive index page and its components to app directory
