--- src/components/archive/ArchiveList/index.tsx
@@ -1,1 +1,0 @@
-export { default } from "./ArchiveList";
